new file:   src/components/CitiesListProps.tsx modified:   src/components/cities-list.tsx modified:   src/pages/main-page/main-page.tsx
